Moved all extra defines to new ux_api.h
